moved install call to ofmt.js
