--- install.js
@@ -1,58 +1,0 @@
-#!/usr/bin/env node
-import {fileURLToPath} from 'url'
-import path from 'path'
-import fs from 'fs'
-
-const __dirname = path.dirname(fileURLToPath(import.meta.url))
-const refDir = process.env.INIT_CWD
-
-// ANCHOR 1. Create symlink to .prettierrc
-const prettierrcFilename = '.prettierrc'
-const prettierrcPath = path.join(__dirname, prettierrcFilename)
-const prettierrcTargetPath = path.join(refDir, prettierrcFilename)
-const prettierrcTargetPathStats = fs.statSync(prettierrcTargetPath, {throwIfNoEntry: false})
-
-if (!prettierrcTargetPathStats) {
-  fs.symlinkSync(prettierrcPath, prettierrcTargetPath, 'file')
-  console.log(`Added new prettier config to your project in "${prettierrcFilename}" file.`)
-}
-
-// ANCHOR 2. Add "eslint.quality.cjs" to extends of eslint config within package.json
-const eslintConfigFilename = 'eslint.quality.cjs'
-const packageJsonPath = path.join(refDir, 'package.json')
-const eslintConfigPath = path.join(__dirname, eslintConfigFilename)
-const eslintConfigRelPath = path.relative(refDir, eslintConfigPath)
-const packageJson = JSON.parse(fs.readFileSync(packageJsonPath, {encoding: 'utf-8'}))
-let updated = false
-
-if (!packageJson.eslintConfig) {
-  packageJson.eslintConfig = {extends: eslintConfigRelPath}
-  updated = true
-  console.log('Added new eslint config to your package.json.')
-}
-
-if (packageJson.eslintConfig && !packageJson.eslintConfig.extends) {
-  packageJson.eslintConfig.extends = eslintConfigRelPath
-  updated = true
-  console.log('Your eslint config in package.json is extended.')
-}
-
-if (packageJson.eslintConfig && packageJson.eslintConfig.extends) {
-  const existingConfig = packageJson.eslintConfig.extends
-
-  if (Array.isArray(existingConfig) && !existingConfig.includes(eslintConfigRelPath)) {
-    packageJson.eslintConfig.extends.push(eslintConfigRelPath)
-    updated = true
-    console.log('Your eslint config in package.json is extended.')
-  }
-
-  if (!Array.isArray(existingConfig) && existingConfig !== eslintConfigRelPath) {
-    packageJson.eslintConfig.extends = [existingConfig, eslintConfigRelPath]
-    updated = true
-    console.log('Your eslint config in package.json is extended.')
-  }
-}
-
-if (updated) {
-  fs.writeFileSync(packageJsonPath, JSON.stringify(packageJson, null, 2))
-}
