+++ package.json
@@ -31,7 +31,7 @@
   "scripts": {
     "lint": "eslint .",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "install": "node install.js"
+    "install": "./bin/ofmt.js install"
   },
   "eslintConfig": {
     "extends": "./eslint.quality.cjs"
